rename sparse/sparse to sparse/array
